refactor(312): change parent folder name and add proper error handling
